added LLM mlp
